# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.bankRequest:
    properties:
      is_kas:
        example: 0
        type: integer
      nama_bank:
        example: BRI
        type: string
      nama_rekening:
        example: PT Kavling Mentaya
        type: string
      no_rekening:
        example: "1234567890"
        type: string
    required:
    - nama_bank
    - nama_rekening
    - no_rekening
    type: object
...
  handlers.marketingRequest:
    properties:
      alamat:
        example: Jl. Contoh No. 1, Sampit
        type: string
      email:
        example: budi@email.com
        type: string
      nama:
        example: Budi Santoso
        type: string
      no_telp:
        example: "08123456789"
        type: string
      persentase_komisi:
        example: 2.5
        type: number
      status:
        example: 1
        type: integer
    required:
    - nama
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@kavling-mentaya.com
    name: API Support
  description: REST API backend untuk sistem manajemen penjualan kavling tanah. Semua
    endpoint (kecuali /auth/login) memerlukan Bearer Token JWT di header Authorization.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Kavling Mentaya API
  version: "1.0"
paths:
...
  /bank:
    get:
      description: Mengambil daftar semua rekening bank dan kas yang aktif
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List rekening bank/kas
      tags:
      - Bank
    post:
      consumes:
      - application/json
      description: Menambahkan rekening bank atau kas tunai baru
      parameters:
      - description: Data bank
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.bankRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tambah rekening bank/kas
      tags:
      - Bank
  /bank/{id}:
    delete:
      description: Menghapus rekening bank berdasarkan ID
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Hapus rekening bank/kas
      tags:
      - Bank
    put:
      consumes:
      - application/json
      description: Mengubah data rekening bank berdasarkan ID
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data update bank
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.bankRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update rekening bank/kas
      tags:
      - Bank
...
  /customers:
    get:
      description: Mengambil daftar semua customer (pembeli)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List customer
      tags:
      - Customer
    post:
      consumes:
      - multipart/form-data
      description: Membuat data customer baru. Mendukung upload file foto KTP dan
        KK (multipart/form-data).
      parameters:
      - description: Nama lengkap
        in: formData
        name: nama
        required: true
        type: string
      - description: No. telepon
        in: formData
        name: no_telp
        type: string
      - description: Nomor KTP
        in: formData
        name: no_ktp
        type: string
      - description: Alamat
        in: formData
        name: alamat
        type: string
      - description: Pekerjaan
        in: formData
        name: pekerjaan
        type: string
      - description: Foto KTP (jpg/png)
        in: formData
        name: foto_ktp
        type: file
      - description: Foto KK (jpg/png)
        in: formData
        name: foto_kk
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tambah customer
      tags:
      - Customer
  /customers/{id}:
    delete:
      description: Menghapus customer berdasarkan ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Hapus customer
      tags:
      - Customer
    get:
      description: Mengambil detail satu customer berdasarkan ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Detail customer
      tags:
      - Customer
    put:
      consumes:
      - multipart/form-data
      description: Mengubah data customer. Mendukung re-upload foto KTP/KK.
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nama lengkap
        in: formData
        name: nama
        type: string
      - description: No. telepon
        in: formData
        name: no_telp
        type: string
      - description: Nomor KTP
        in: formData
        name: no_ktp
        type: string
      - description: Alamat
        in: formData
        name: alamat
        type: string
      - description: Pekerjaan
        in: formData
        name: pekerjaan
        type: string
      - description: Foto KTP baru
        in: formData
        name: foto_ktp
        type: file
      - description: Foto KK baru
        in: formData
        name: foto_kk
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update customer
      tags:
      - Customer
...
  /marketing:
    get:
      description: Mengambil daftar semua marketing beserta jumlah transaksi OPEN
        dan CLOSED
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List marketing
      tags:
      - Marketing
    post:
      consumes:
      - application/json
      description: Menambahkan data marketing (sales) baru
      parameters:
      - description: Data marketing
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.marketingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tambah marketing
      tags:
      - Marketing
  /marketing/{id}:
    delete:
      description: Menghapus marketing. Ditolak jika marketing sudah memiliki transaksi.
      parameters:
      - description: Marketing ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Sudah ada transaksi
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Hapus marketing
      tags:
      - Marketing
    get:
      description: Mengambil detail satu marketing beserta jumlah transaksi OPEN dan
        CLOSED
      parameters:
      - description: Marketing ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Detail marketing
      tags:
      - Marketing
    put:
      consumes:
      - application/json
      description: Mengubah data marketing berdasarkan ID
      parameters:
      - description: Marketing ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data update marketing
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.marketingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update marketing
      tags:
      - Marketing
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}. Token diperoleh
      dari endpoint POST /api/auth/login.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"